Moving minor items to better locations.
